requirements.txt
torch
transformers>=4.36.0,<4.46.0
datasets>=2.14.0
peft>=0.6.0
trl>=0.11.0,<0.12.0
wandb
huggingface_hub>=0.16.0
python-dotenv
accelerate>=0.24.0
